# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto Controle de filmes SENAI Jandira - Matheus Henry
  description: |-
   API destinada a controle de filmes para o projeto bem legal do SENAI
  termsOfService: http://swagger.io/terms/
  contact:  
    email: 09matheushenry@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://
paths:
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo Filme
      description: End point para enviar os dados do filme a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo Filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '415':
          description: Invalid Content-Type
        '500':
          description: Error internal Server (Controller / model)
    get:
      tags:
        - Filme
      summary: Lista todos os Filmes
      description: Listagem de todos os Filme registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'          
           
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Controller/Model)
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Busca um Filme pelo ID
      description: Retorna um Filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do filme
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found
        '500':
          description: Error internal server (Controller/Model)
    put:
      tags:
        - Filme
      summary: Atualiza um filme pelo id
      description: Buscar e atualizasr um Filme pelo id
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um Filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found
        '415':
          description: Invalid content type
        '500':
          description: Error Internal Server (Controller/Model)
    delete:
      tags:
        - Filme
      summary: Delete Filme
      description: Excluir um Filme
      parameters:
        - name: id
          in: path
          description: ID of pet to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
        '500': Error Internal Server (Controller/Model)
components:
  schemas:
    Filme:
      required:
        - nome
        - sinopse
        - data_lancamento
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sinopse:
          type: string
        data_lancamento:
          type: string
        duracao:
          type: string
        capa:
          type: string
        valor:
          type: integer
        avaliacao:
          type: integer
  requestBodies:
    Filme:
      description: Objeto que deve ser enviado no Body das requisicoes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Filme'
    